+++ app/controllers/profiles_controller.rb
@@ -4,6 +4,9 @@
 
 class ProfilesController < ApplicationController
   before_filter :authenticate_user!, :except => ['show']
+  before_filter -> { @css_framework = :bootstrap }, only: [:show, :edit]
+ 
+  layout ->(c) { request.format == :mobile ? "application" : "with_header_with_footer" }, only: [:show, :edit]
 
   respond_to :html, :except => [:show]
   respond_to :js, :only => :update