+++ app/controllers/services_controller.rb
@@ -8,6 +8,9 @@ class ServicesController < ApplicationController
   skip_before_filter :verify_authenticity_token, :only => :create
   before_filter :authenticate_user!
   before_filter :abort_if_already_authorized, :abort_if_read_only_access, :only => :create
+  before_filter -> { @css_framework = :bootstrap }, only: [:index]
+
+  layout ->(c) { request.format == :mobile ? "application" : "with_header_with_footer" }, only: [:index]
 
   respond_to :html
   respond_to :json, :only => :inviter